fix: cast tensors to float32 before moving to MPS device
MPS does not support float64; dict_to_device now calls .float() before .to(device).

+++ direct/utils/__init__.py
@@ -172,7 +172,7 @@ def dict_to_device(
     """
     if keys is None:
         keys = data.keys()
-    return {k: v.to(device) if isinstance(v, torch.Tensor) else v for k, v in data.items() if k in keys}
+    return {k: v.float().to(device) if isinstance(v, torch.Tensor) else v for k, v in data.items() if k in keys}
 
 
 def detach_dict(data: Dict[str, torch.Tensor], keys: Optional[Union[List, Tuple, KeysView]] = None) -> Dict:
